Sergeev

See merge request mobile/saver-android!21

+++ app/src/main/java/digital/fact/saver/presentation/fragments/plan/RefactorCurrentPlanFragment.kt
@@ -7,17 +7,18 @@ import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
-import com.google.android.material.progressindicator.BaseProgressIndicator
 import com.prolificinteractive.materialcalendarview.CalendarDay
 import digital.fact.saver.R
 import digital.fact.saver.data.database.dto.PlanTable
 import digital.fact.saver.databinding.FragmentRefactorCurrentPlanBinding
+import digital.fact.saver.presentation.dialogs.ConfirmDeleteDialog
 import digital.fact.saver.presentation.viewmodels.PlansViewModel
 import digital.fact.saver.utils.*
 import digital.fact.saver.utils.calandarView.CurrentDayDecoratorInRange
@@ -30,7 +31,6 @@ import org.threeten.bp.Instant
 import org.threeten.bp.LocalDate
 import org.threeten.bp.ZoneId
 import java.util.*
-
 
 class RefactorCurrentPlanFragment  : Fragment() {
 
@@ -71,10 +71,10 @@ class RefactorCurrentPlanFragment  : Fragment() {
                     if (item.id == id) {
                         this.plan = item
                         binding.textViewSumLogo.text =
-                                if (item.type == PlanTable.PlanType.SPENDING.value)
+                                if (item.type == PlanTable.PlanType.EXPENSES.value)
                                     resources.getString(R.string.plan_spending_2)
                                 else resources.getString(R.string.plan_income_2)
-                        binding.toolbar.subtitle = if (item.type == PlanTable.PlanType.SPENDING.value)
+                        binding.toolbar.subtitle = if (item.type == PlanTable.PlanType.EXPENSES.value)
                             resources.getString(R.string.spend)
                         else resources.getString(R.string.income)
                         val date = Date(item.planning_date)
@@ -96,9 +96,15 @@ class RefactorCurrentPlanFragment  : Fragment() {
             when (item?.itemId) {
                 R.id.delete_plan -> {
                     plan?.let { currentPlan ->
-                        plansVM.deletePlan(currentPlan).observe(viewLifecycleOwner, {
-                            navC.popBackStack()
-                        })
+                        ConfirmDeleteDialog(title = getString(R.string.will_do_delete),
+                                description = getString(R.string.you_delete_plan_from_list),
+                                onSliderFinishedListener = {
+                                    plansVM.deletePlan(currentPlan).observe(viewLifecycleOwner, {
+                                        Toast.makeText(requireContext(), getString(R.string.deleted), Toast.LENGTH_SHORT).show()
+                                        navC.popBackStack()
+                                    })
+                                }
+                        ).show(childFragmentManager, "confirm-delete-dialog")
                     }
                 }
             }
@@ -128,7 +134,7 @@ class RefactorCurrentPlanFragment  : Fragment() {
             }
         })
 
-        binding.buttonAddPlan.setOnClickListener { _ ->
+        binding.buttonAddPlan.setOnClickListener {
             this.plan?.let { planCurrent ->
                 val date = DateTimeUtils.toSqlDate(binding.calendar.selectedDate?.date).time
                 val sum: Long = (round(binding.editTextSum.text.toString().toDouble(), 2) * 100).toLong()
