+++ app/src/main/java/digital/fact/saver/presentation/fragments/plan/RefactorPlanFragment.kt
@@ -17,6 +17,7 @@ import androidx.navigation.fragment.findNavController
 import digital.fact.saver.R
 import digital.fact.saver.data.database.dto.PlanTable
 import digital.fact.saver.databinding.FragmentPlanCompletedRefactorBinding
+import digital.fact.saver.presentation.dialogs.ConfirmDeleteDialog
 import digital.fact.saver.presentation.viewmodels.OperationsViewModel
 import digital.fact.saver.presentation.viewmodels.PlansViewModel
 import digital.fact.saver.utils.*
@@ -81,15 +82,15 @@ class RefactorPlanFragment : Fragment() {
                                 }
                             }
                             binding.textViewSumLogo.text =
-                                    if (plan.type == PlanTable.PlanType.SPENDING.value)
+                                    if (plan.type == PlanTable.PlanType.EXPENSES.value)
                                         resources.getString(R.string.plan_spending_2)
                                     else resources.getString(R.string.plan_income_2)
                             binding.textViewFactLogo.text =
-                                    if (plan.type == PlanTable.PlanType.SPENDING.value)
+                                    if (plan.type == PlanTable.PlanType.EXPENSES.value)
                                         resources.getString(R.string.fact_spent)
                                     else resources.getString(R.string.fact_income)
                             binding.toolbar.subtitle =
-                                    if (plan.type == PlanTable.PlanType.SPENDING.value)
+                                    if (plan.type == PlanTable.PlanType.EXPENSES.value)
                                         resources.getString(R.string.spend)
                                     else resources.getString(R.string.income)
                             binding.editTextDescription.setText(plan.name)
@@ -112,26 +113,41 @@ class RefactorPlanFragment : Fragment() {
             when (item?.itemId) {
                 R.id.plan_refactor_done_in_range_delete -> {
                     plan?.let { currentPlan ->
-                        plansVM.deletePlan(currentPlan).observe(viewLifecycleOwner, {
-                            Toast.makeText(requireContext(), getString(R.string.deleted), Toast.LENGTH_SHORT).show()
-                            navC.popBackStack()
-                        })
+                        ConfirmDeleteDialog(title = getString(R.string.will_do_delete),
+                                description = getString(R.string.you_delete_plan_from_list),
+                                onSliderFinishedListener = {
+                                    plansVM.deletePlan(currentPlan).observe(viewLifecycleOwner, {
+                                        Toast.makeText(requireContext(), getString(R.string.deleted), Toast.LENGTH_SHORT).show()
+                                        navC.popBackStack()
+                                    })
+                                }
+                        ).show(childFragmentManager, "confirm-delete-dialog")
                     }
                 }
                 R.id.plan_refactor_done_in_range_reset -> {
                     plan?.let { currentPlan ->
-                        val updatePlan = PlanTable(
-                                id = currentPlan.id,
-                                type = currentPlan.type,
-                                sum = currentPlan.sum,
-                                name = currentPlan.name,
-                                operation_id = 0,
-                                planning_date = currentPlan.planning_date
-                        )
-                        plansVM.updatePlan(updatePlan).observe(viewLifecycleOwner, {
-                            Toast.makeText(requireContext(), resources.getString(R.string.reseted), Toast.LENGTH_SHORT).show()
-                            navC.popBackStack()
-                        })
+
+
+
+
+                        ConfirmDeleteDialog(title = getString(R.string.will_do_reset),
+                                description = getString(R.string.you_will_reset_plan),
+                                onSliderFinishedListener = {
+                                    val updatePlan = PlanTable(
+                                            id = currentPlan.id,
+                                            type = currentPlan.type,
+                                            sum = currentPlan.sum,
+                                            name = currentPlan.name,
+                                            operation_id = 0,
+                                            planning_date = currentPlan.planning_date
+                                    )
+                                    plansVM.updatePlan(updatePlan).observe(viewLifecycleOwner, {
+                                        Toast.makeText(requireContext(), resources.getString(R.string.reseted), Toast.LENGTH_SHORT).show()
+                                        navC.popBackStack()
+                                    })
+                                }
+                        ).show(childFragmentManager, "confirm-delete-dialog")
+
                     }
                 }
                 R.id.plan_refactor_done_in_range_show_history -> {
@@ -168,7 +184,7 @@ class RefactorPlanFragment : Fragment() {
             }
         })
 
-        binding.buttonAddPlan.setOnClickListener { _ ->
+        binding.buttonAddPlan.setOnClickListener {
             this.plan?.let { planCurrent ->
                 val sum: Long =
                         (round(binding.editTextSum.text.toString().toDouble(), 2) * 100).toLong()
